+++ website/next.config.mjs
@@ -7,6 +7,14 @@ const nextConfig = {
     unoptimized: true,
   },
   trailingSlash: true,
+  webpack: (config) => {
+    config.module.rules.push({
+      test: /\.md$/,
+      use: "raw-loader",
+    });
+
+    return config;
+  },
 };
 
 export default nextConfig;